pyproject.toml
[project]
name = "metasploit-mcp-server"
version = "0.1.0"
description = "MCP server for interacting with Metasploit Framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "mcp[cli]>=1.3.0",
    "pymetasploit3>=1.0.6",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools", "utils"]
